move LanguageCodes under core
