+++ core/lang.ts
